Move tests to __tests__ directory
