+++ tests/CalendarMonth.spec.jsx
@@ -1,18 +1,19 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import moment from 'moment';
-import { CalendarMonth } from './CalendarMonth';
+import { CalendarMonth } from '../client/src/components/CalendarMonth/CalendarMonth';
 
 describe('Calender Month component', () => {
   let calendarMonth;
   beforeEach(() => {
-    const dayStatuses = [];
-    dayStatuses.length = moment('2020-03').daysInMonth();
-    dayStatuses.fill('available');
+    const reservedDays = [];
+    reservedDays.length = moment('2020-03').daysInMonth();
+    reservedDays.fill('available');
     calendarMonth = shallow(
       <CalendarMonth
         month="2020-03"
-        dayStatuses={dayStatuses}
+        reservedDays={reservedDays}
+        position={0}
       />,
     );
   });
